requirements.txt
jax
jaxlib
numpy
redis
chromadb
pyyaml
pytest
